control_tower/requirements.txt
Flask==0.12
Flask_Login==0.4.0
flask_paginate==0.4.5
Flask_Principal==0.4.0
Flask_SQLAlchemy==2.2
Flask_WTF==0.14.2
WTForms==2.1
python_bcrypt==0.3.2
